Way better / cleaner implementation
