+++ ui/src/createTelemetryReporter.ts
@@ -6,9 +6,6 @@
 import * as process from 'process';
 import * as vscode from 'vscode';
 import TelemetryReporter from 'vscode-extension-telemetry';
-import { IExperimentationTelemetry } from 'vscode-tas-client';
-import { IActionContext } from '..';
-import { registerTelemetryHandler } from './callWithTelemetryAndErrorHandling';
 import { DebugReporter } from './DebugReporter';
 import { getPackageInfo } from './getPackageInfo';
 
@@ -17,9 +14,8 @@ const debugTelemetryEnabled: boolean = !/^(false|0)?$/i.test(process.env.DEBUGTE
 // tslint:disable-next-line:strict-boolean-expressions
 const debugTelemetryVerbose: boolean = /^(verbose|v)$/i.test(process.env.DEBUGTELEMETRY || '');
 
-export interface IInternalTelemetryReporter extends IExperimentationTelemetry {
+export interface IInternalTelemetryReporter {
     sendTelemetryErrorEvent(eventName: string, properties?: { [key: string]: string | undefined }, measurements?: { [key: string]: number | undefined }, errorProps?: string[]): void;
-    handleTelemetry(context: IActionContext): void;
 }
 
 export function createTelemetryReporter(ctx: vscode.ExtensionContext): IInternalTelemetryReporter {
@@ -31,50 +27,13 @@ export function createTelemetryReporter(ctx: vscode.ExtensionContext): IInternal
         console.warn(`${extensionName}: DEBUGTELEMETRY mode enabled (${debugTelemetryVerbose ? 'verbose' : 'quiet'}) - not sending telemetry`);
         newReporter = new DebugReporter(extensionName, extensionVersion, debugTelemetryVerbose);
     } else {
-        const reporter: InternalTelemetryReporter = new InternalTelemetryReporter(extensionName, extensionVersion, aiKey);
+        const reporter: TelemetryReporter = new TelemetryReporter(extensionName, extensionVersion, aiKey);
         ctx.subscriptions.push(reporter);
         newReporter = reporter;
     }
-
-    ctx.subscriptions.push(registerTelemetryHandler((context: IActionContext) => { newReporter.handleTelemetry(context); }));
 
     // Send an event with some general info
     newReporter.sendTelemetryErrorEvent('info', { isActivationEvent: 'true', product: vscode.env.appName, language: vscode.env.language }, undefined, []);
 
     return newReporter;
-}
-
-class InternalTelemetryReporter extends TelemetryReporter implements IInternalTelemetryReporter {
-    private readonly sharedProperties: { [key: string]: string } = {};
-
-    /**
-     * Implements `postEvent` for `IExperimentationTelemetry`.
-     * @param eventName The name of the event
-     * @param props The properties to attach to the event
-     */
-    public postEvent(eventName: string, props: Map<string, string>): void {
-        const properties: { [key: string]: string } = {};
-
-        for (const key of props.keys()) {
-            properties[key] = <string>props.get(key);
-        }
-
-        this.sendTelemetryErrorEvent(eventName, properties);
-    }
-
-    /**
-     * Implements `setSharedProperty` for `IExperimentationTelemetry`
-     * @param name The name of the property
-     * @param value The value of the property
-     */
-    public setSharedProperty(name: string, value: string): void {
-        this.sharedProperties[name] = value;
-    }
-
-    public handleTelemetry(context: IActionContext): void {
-        context.telemetry.properties = {
-            ...context.telemetry.properties,
-            ...this.sharedProperties
-        };
-    }
 }
